requirements.txt
pyserial==2.7
tornado==3.2
RPi.GPIO==0.5.5
Pillow==2.5.1

